+++ src/core/datastructures/camera/orthographiccamera.cpp
@@ -39,7 +39,7 @@
 
 namespace inviwo {
 
-OrthographicCamera::OrthographicCamera(vec3 lookFrom, vec3 lookTo, vec3 lookUp, float nearPlane,
+OrthographicCamera::OrthographicCamera(dvec3 lookFrom, dvec3 lookTo, dvec3 lookUp, float nearPlane,
                                        float farPlane, float aspectRatio, float width)
     : Camera(lookFrom, lookTo, lookUp, nearPlane, farPlane, aspectRatio), width_{width} {}
 
@@ -106,15 +106,15 @@ bool OrthographicCamera::equal(const Camera& other) const {
     }
 }
 
-mat4 OrthographicCamera::calculateProjectionMatrix() const {
-    const float halfWidth = 0.5f * width_;
-    const float halfHeight = halfWidth / aspectRatio_;
-    return glm::ortho(-halfWidth, +halfWidth, -halfHeight, +halfHeight, nearPlaneDist_,
-                      farPlaneDist_);
+dmat4 OrthographicCamera::calculateProjectionMatrix() const {
+    const double halfWidth = 0.5 * width_;
+    const double halfHeight = halfWidth / aspectRatio_;
+    return glm::ortho(-halfWidth, +halfWidth, -halfHeight, +halfHeight,
+                      static_cast<double>(nearPlaneDist_), static_cast<double>(farPlaneDist_));
 }
 
-vec4 OrthographicCamera::getClipPosFromNormalizedDeviceCoords(const vec3& ndcCoords) const {
-    return vec4{ndcCoords, 1.0f};
+dvec4 OrthographicCamera::getClipPosFromNormalizedDeviceCoords(const dvec3& ndcCoords) const {
+    return dvec4{ndcCoords, 1.0};
 }
 
 void OrthographicCamera::serialize(Serializer& s) const {
